Add short option accepting a value without a space
This commits adds support for short options with no space/equal sign for
_all_ types. The tests are changed accordingly. Even though supporting
this is quite controversial (many cases result in questionable behavior,
e.g. -timeout is read as -t imeout, if t is a short string option), it
is the default to avoid breaking code. But you can enable it only for
short numeric options via noSpaceOnlyForShortNumericOptions.
Further we throw an exception if a Boolean option is given with a value.
The commit also updates documentation and indentation to ease reading. A
faulty $(UL ) is removed.